# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Gin API with Swagger documentation.",
        "title": "ECOM",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/category": {
            "get": {
                "description": "Admin can see the listed  category in ecommerce website",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Category"
                ],
                "summary": "get all category list by admin",
                "responses": {}
            },
            "post": {
                "description": "Add a new category with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Category"
                ],
                "summary": "Add a new category",
                "operationId": "add-category",
                "parameters": [
                    {
                        "description": "Category Add",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.CategoryData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Duplicate found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/category/{ID}": {
            "put": {
                "description": "Admin can edit the listed  category in ecommerce website",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Category"
                ],
                "summary": "Admin can edit the category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category Edit",
                        "name": "Form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.CategoryData"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Deletes a category by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Category"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "Admin can Block the listed  category in ecommerce website",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Category"
                ],
                "summary": "Admin can block the category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/coupon": {
            "get": {
                "description": "Admin side list all the coupons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Coupons"
                ],
                "responses": {
                    "200": {
                        "description": "Coupons",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Admin can add the coupon in ecommerce website",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Coupons"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coupon Code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coupon Amount",
                        "name": "amount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit Amount",
                        "name": "limit",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon added",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/coupon/{ID}": {
            "delete": {
                "description": "Admin side delete a coupon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Coupons"
                ],
                "summary": "Admin can delete a coupon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coupon ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/offer/{ID}": {
            "post": {
                "description": "Product Offer Adding",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Offer"
                ],
                "summary": "Offer Adding",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Offer",
                        "name": "offer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.OfferProductData"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Product Offer Deleting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Offer"
                ],
                "summary": "Product Offer Deleting",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/order/update/{ID}": {
            "patch": {
                "description": "Update the status of an order by ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Order"
                ],
                "summary": "Update Order Status",
                "operationId": "update-order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New status",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "database.CategoryData": {
            "type": "object",
            "properties": {
                "catagory": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "database.OfferProductData": {
            "type": "object",
            "properties": {
                "expirey": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                }
            }
        }
    }
}